+++ tests/test_api.lua
@@ -12,7 +12,6 @@ function T.setup()
     RecipeBookDB.characters = {}
     RecipeBookDB.minimap = { hide = false }
     RecipeBookDB.maxPhase = 5
-    RecipeBookDB.currentPhase = 1
     RecipeBookCharDB = RecipeBookCharDB or {}
     RecipeBookCharDB.professionSkill = {}
     RecipeBook:GetMyCharData()