+++ tests/test_tooltip_status.lua
@@ -22,7 +22,6 @@ function T.setup()
     RecipeBookDB.ignoredCharacters = nil
     RecipeBookDB.minimap = { hide = false }
     RecipeBookDB.maxPhase = 5
-    RecipeBookDB.currentPhase = 1
     RecipeBookCharDB = RecipeBookCharDB or {}
     RecipeBookCharDB.hideKnown = false
     RecipeBookCharDB.collapsedSources = {}